pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linkedin-gmail-jobs-hub"
version = "0.1.0"
description = "Phase 1 API to centralize LinkedIn job emails from Gmail"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy>=2.0.40",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.9.0",
  "python-dotenv>=1.1.0",
  "beautifulsoup4>=4.13.0",
  "lxml>=5.3.0",
  "google-api-python-client>=2.166.0",
  "google-auth>=2.38.0",
  "google-auth-oauthlib>=1.2.1",
  "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=6.1.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
